+++ includes/class-tmsm-woocommerce-customadmin.php
@@ -69,7 +69,7 @@ class Tmsm_Woocommerce_Customadmin {
 	public function __construct() {
 
 		$this->plugin_name = 'tmsm-woocommerce-customadmin';
-		$this->version     = '1.0.9';
+		$this->version     = '1.0.10';
 
 		$this->load_dependencies();
 		$this->set_locale();