+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 * Sort order actions: view in first position
 * Color for status: processed
 * Color for order actions buttons: processing, complete, processed
+* Rename order statuses
+* Update i18n
 
 ### 1.0.9: January 9th, 2018
 * New status "processed"